+++ maven-resolver-provider/src/main/java/org/apache/maven/repository/internal/MavenArtifactRelocationSource.java
@@ -32,9 +32,9 @@ public interface MavenArtifactRelocationSource {
     /**
      * Returns {@link Artifact} instance where to relocate to, or {@code null}.
      *
-     * @param session  The session, never {@code null}.
+     * @param session The session, never {@code null}.
      * @param request The artifact descriptor request, never {@code null}.
-     * @param model    The artifact model, never {@code null}.
+     * @param model The artifact model, never {@code null}.
      * @return The {@link Artifact} to relocate to, or {@code null} if no relocation wanted.
      */
     Artifact relocatedTarget(RepositorySystemSession session, ArtifactDescriptorRequest request, Model model);
